Ensure a full request/response happens during conntrack offset guessing
We have been seeing sporadic conntrack test failures in CI, and sporadic
failures to initialize the eBPF conntracker internally.

For background, offset guessing for a single field works within a loop where the
same event is repeadly generated, and an ebpf program reads the contents
of a function param at a candidate offset which is advanced with every
iteration of the loop until the contents at that candidate offset match
the contents at an event.

```
while guessing:
    ev := eventGenerator.Generate() // generates a UDP connection, populates guessed with the current guessed offsets
    if guessed[field] == ev[field]
      field = next field to guess
    else:
      candidate offset for field ++
```

This PR may help solve the problem by ensuring that during conntrack
offset guessing each Generate() event ensures a UDP connection is
established all the way (i.e. writes and then reads). Before this
change, the Generate() event only ran a net.Dial, which might not have
reliably caused a conntrack entry to be established.