pkg/trace/api: fix double-close deadlock (#20761)
The onCloseConn does not guard against multiple invocations of Close,
meaning the side-effects can be executed multiple times. This can result
in a deadlock on the shutdown routine, resulting in the agent failing to
shut down.

This change introduces a sync.Once element to ensure onCloseConn only
executes its callback once.