Migrate secret pkg to component
This change creates a `secrets.Component` that is implemented by an instance instead of global state. The instance class is `secretsimpl.secretResolver`

As part of this change, the agent's usage of secrets makes these changes:

* remove the methods config.NewAgentParamsWith{out}Secrets
  - instead we provide a secrets.NewEnabledParams() constructor
  - if not used, secrets is disabled by default
* remove the bool coreParams.ConfigLoadSecrets()
  - instead secretsParams.Enable is used
  - pass secretsParams as an fx dependency to the secretResolver
* change helper SetupConfigWithWarnings -> SetupConfigForTest
* remove the "secrets" build tag
* remove NewServerlessConfig's reference to WithConfigLoadSecret
  - serverless shouldn't use secrets, didn't have a build tag before
* remove configLoadSecrets from comp/core/config/params.go
  - let the secretsParam be the sole point of configuration instead
* stop almost all usage from the secrets component's global state

One hack is being added. The diagnose callbacks don't have access to components, so we need to get a reference to a global secrets component. We do this using a `GetInstance()` method, which returns a singleton instance protected by a mutex. This hack is only needed because changing the signature of diagnose is a larger project than can be done by this PR.