http2: Extract conn tuple from http2_stream_t
We can remove the instance of conn_tuple_t from http2_stream_t, and by that minimize the size of
http2_in_flight map. The conn_tuple_t field in http2_stream_t is redundant as it exists in the key
of the map. Alongside we removing the duplication, we introduce a new struct called http2_event_t
which composes both conn_tuple_t and http2_stream_t struct, and send http2_event_t to the user mode.
By that, we reduce the size of http2_in_flight without affecting the events sent to the usermode.