http2: Bump the value of HTTP2_FRAMES_PER_TAIL_CALL to 9 (#20945)
With a simple change to the main loop of socket__http2_frames_parser, we can increase the value of HTTP2_FRAMES_PER_TAIL_CALL to 9.
The change - is to move away from the loop, a redundant reinitialization of the http2_ctx, which does not change (at least
the parts we're reseting) during the iterations. The change allow us to bump the number of frames to parse in a single tail call
to 9 (instead of 4 today). This change means we can parse up to 29*9 (261) frames instead of 29*4 (116) in total.