[USM] http2 dynamic table cleanup (#20913)
* usm: map cleaner: Allow pre processing and post processing for the cleanup.

The new methods should allow us to prepare for a cleanup cycle (like creating a list for deletion, or to synchronize elements)

* usm: http2: Change http2_dynamic_table to hash map

LRU contributed to flakiness in our kernel code. We'll introduce a cleanup in the user mode to avoid filling the map forever

* usm: http2: Added terminated events for http2 connections

A preliminary step for cleanup of http2_dynamic_table. We are laying the infrastructure
for the actual cleanup. In the current commit we are sending from the kernel to the user mode
the closed connections.

* usm: http2: Rename mapCleaner to http2InFlightMapCleaner

* usm: http2: Save closed connections in the user mode

* usm: http2: Enable http2_dynamic_table cleanup

* usm: http2: Introduce configuration to control the interval of the cleanup

* usm: http2: Added a UT to check the cleanup

* Fixed CR

* usm: http2: Rename configuration value to make it more readable

* usm: http2: Remove redundant creation of by-value mutex

* usm: http2: Remove redundant mutex

The mutex is not needed as we have a single go-routine working on the map

* usm: http2: Fixed wrong condition

* usm: http2: Remove reudndant creations

* usm: http2: Remove intermediate array, and clean map

* Fixed CR

* Undo import change

* fixed CR