fix: appsec-node serverless integration test failure (#20798)
Somehow this is not registering a trace anymore. For some reason I fail to understand, when instrumentation telemetry is enabled (which is te default behavior when `DD_SERVERLESS_APPSEC_ENABLED` is `true`-ish), the `appsec-node` function no longer produces a trace in a way that is captured by the `recorder-extension`. I have however been able to validate that the trace is actually still created (in a run with `DD_LOG_LEVEL=debug`).

In order to fix this, explicitly disabled instrumentation telemetry so the test continues to work correctly. The actual interaction needs to be investigated later/separately, as it is not well understood today why either the Node tracer library or the recorder extension behaves differently in this particular scenario.

---

Additionally, added extra logging into the recorder extension (as I was attempting to understand the issue), and improved the log normalization to also neutralize the AWS account ID.

---

Finally, using feature environments to reduce the amount of deployed functions (in the serverless stacks) depending on the test suite being run.