usm: http2: Simplify methods to reduce program's complexity
The change is maintaining the functionality of the code, but does a simple trick
to spare if-clauses which added complexity to the bpf program, and lead us
to suffer from -EBIG error when verified.
The idea of the trick is to override the entries in the headers_to_process array
but advancing the counter only if the entry is relevant.

In the case where we've found all of our interesting headers, we cannot override
the last entry in the array, as there is a condition in filter_relevant_headers
which sets headers_to_process as NULL, in such a case, so we won't override
by mistake any interesting entry