Add more synchronization to DNS tests
We saw a few tests fail with the following error:

```
=== FAIL: pkg/network/tracer TestTracerSuite/prebuilt/TestDNSStatsWithNAT (3.76s)
    tracer_test.go:1043:
        	Error Trace:	/go/src/github.com/DataDog/datadog-agent/pkg/network/tracer/tracer_test.go:1043
        	            				/go/src/github.com/DataDog/datadog-agent/pkg/network/tracer/tracer_linux_test.go:1380
        	Error:      	Should be true
        	Test:       	TestTracerSuite/prebuilt/TestDNSStatsWithNAT

```

My theory is that the issue is that testDNSStats calls getConnections,
which itself calls reverseDns#GetDNSStats().

Under the hood, reverseDNS#GetDNSStats() is polling a raw socket for DNS
packets.
My theory is that reverseDNS hasn't reliably polled the latest DNS stat by the time the test calls getConnections().

This PR fixes the problem by adding a synchronization point where the
test waits for the reverseDNS to see the domain. This is done a bit
crudely with a test-only method that waits up to 5 seconds for the
domain to show up.