usm: http2: Added initial prototype of user mode dynamic table
The new user mode dynamic table is meant to store HTTP2 paths in a LRU datastore
and to associate stream with their appropriate paths/

The new module is planned to get all paths from the kernel (via perf event)
and to store it in a user-mode LRU datastore, where the key is the connection
tuple and the value index in the dynamic table, and the value in a pointer
to string intern value. Using the string internning we're able to reduce
redundant duplications and spare memory.