Add logging in testDNSStats
TestDNSStatsWithNAT is still flaky, I'm noticing that many of the failures are on prebuilt, and the CO-RE
failures have the suspicious 'clang timed out error'. I'm wondering if
this is an issue of the netlink conntracker being behind the real
tracer. These logs will help verify. If all the failures we see are on
prebuilt, we can either skip this test with netlink conntrack,  or add
some synchronization