Move Setenv to the package init
Forwarder health is now started before we initialize Python, which
creates a possibility for a race between the Setenv call in the Python
init code and getenv used in the DNS resolver.

This patch moves the Setenv call to a package init function instead,
which would hopefully complete before any network connections are
initiated.