Config setting secret_refresh_interval for secrets component
Change ResolveWithCallback to RegisterResolveCallback. This method registers the configuration text, origin, and callback within the resolver, to be reused when Refresh() is called. Only one set of arguments can be registered at a time.

Using the config setting `secret_refresh_interval` allows secrets to be refreshed each given number of seconds. When refreshed, this will invoke the callback passed to the registration.