Fix slice aliasing bug in secrets resolver
Before this change, a path to a secret (like `some.encoded.0.data` in `testConfSibling` in secrets_test.go) could be accidentally modified after being registered. This is because the walker reuses the same slice value while traversing a configuration. By cloning the slice before registering it, we avoid this aliasing bug.