serverless/appsec: fix handling of ALB headers & query string
When an AWS Application Load Balancer forwards a request to Lambda, it
formats headers and query string parameters either as `map[string]string`
in the `Headers` and `QueryStringParameters` properties (this is the
default behavior), or as `map[string][]string` in `MultiValueHeaders`
and `MultiValueQueryStringParameters` (when explicitly enabled by
setting the appropriate target group attribute[^1]).

We were previously only considering the multi-valued entries, since
other event types (such as API Gateway) provide values in both formats
all the time. This change uses the multi-valued map if present, and
falls back to the single-valued map otherwise.

[^1]: https://docs.aws.amazon.com/elasticloadbalancing/latest/application/lambda-functions.html#enable-multi-value-headers