usm: socket filter: Abort if protocol is encrypted
In the context of a socket filter we're not able to decode encrypted traffic.
Thus, we should abort early.
The only exception for this new rule is TCP termination, meaning, if we're identifying
a TCP termination, we should still process it in our probes. This is a fallback for missing
TLS termination and ensure we're cleaning resources.