fix empty platform information
If github.com/shirou/gopsutil/v3 is used and a [/host]/etc/redhat-release file/dir exists but is empty, then it returns empty platform information.

This affects helm-chart users who have kernel header downloading turned on, because it always mounts this file and k8s will create it even if it doesn't exist.