reduce the number of threads for iot-agent packaging on 32bits arch
We can only address 32bits on that architecture, which means we have less than
4GB of addressable memory, and bumping it makes no difference as it can't be
used.
In order to ensure that our packaging job succeeds we need to reduce its memory
demand, which we can do by reducing the number of threads used to compress.
Since the iot-agent is small, it shouldn't make much difference in term of the
job duration