omnibus: iot-agent: use the same compression configuration for armhf builds
dpkg-deb is built as a 32 bits library, causing it not to be able to address
more than 4GB of memory.
In addition, the version we use doesn't seem to correctly handle memory limits,
or dpkg-deb overrides the optios we provide to XZ. This results in the package
compression erroring out with an allocation failure.
Setting the compression level to anything above level 6 resulted in failures
during my tests.