Adding a config command to the trace CLI
The trace agent was the only agent process not offering a way to query
its runtime configuration.
The core-agent will soon need to query the runtime configuration of each
process. This is prep work for this.

We inject the new endpoint from the comp side of the trace agent to
avoid linking pkg/config into pkg/trace.