usm: http2: Add original index field
The new field exists in two structs - http2_header_t and dynamic_table_entry_t.
In dynamic_table_entry_t the purpose is to cache the original index and to know
if the future calls whether the value represents a path, a method or a status code.
In http2_header_t the purpose is to populate the value for its first appearance
in a new dynamic header case.

Currently the only use-case is for paths. Status codes will be added in a later commit"