usm: http2: Support arbitrary status codes
We're supporting the case of a status code being passed as a literal header
instead of a fully indexed entry in the static table.

The PR creates a new struct status_code_t which covers the 2 options we have:
1. indexed_value - represents a fully indexed in the static table
2. literal - so we introduce is_huffman_encoded flag, and a buffer to save the raw value.

In the user mode we're converting the correct status code based on our options.