usm: http2: Add user mode consumer to the perf buffer
The dynamic table management is now able to consume the dynamic values send from the
the eBPF context (socket filter or uprobes). We use a datastore which is a map of cyclic
maps, which maintains a single state per connection tuple, and able to guarantee the
overall size of the datastore does not exceeds a predefined limit. The cyclic map
automatically evicts oldest item when the map is full.