Add new E2E tests
This is a preliminary PR, not ready for checkin yet.

Attempts to replicate the behavior of copying precompiled test executables
to the host, and executing them there.

Includes the ability to set up IIS (a prerequisite for the tests) on windows hosts.

As currently coded, will only work when both the tester and the target are windows hosts.
Clearly needs updating to work with linux testers, as that's the target environment.