fix(serverless/appsec): goroutine leak in TokenTicker tests
The tests for `TokenTicker` rely on a chanel to synchronize with token
bucket updates (each update sends one message to that channel). There
was an issue when closing those test tickers, where it is possible the
ticker goroutine may process the closing of the ticker channel before
the stop signal; leading to one last message being sent to the
synchronization channel. This resulted in the bucket updater goroutine
blocking on the channel send indefinitely, as this last send was not
paired with a receive on the other side.

This could result in those tests leaking a modest amount of goroutines,
affecting the goroutine scheduler's performance. When running tests with
`-count`, this could reliably cause tests to time out, once enough
leaked goroutines have accumulated that the scheduler wastes most of its
time on these.