Add host/container APM instrumentation settings
These settings were previously specified in a separate configuration
file that only the injector processes would read. This was done to
accelerate development in order to get APM instrumentation released in
public beta.

Now that we are looking to make APM instrumentation generally available,
it makes sense to unify this configuration with the rest of the Agent
configuration in the datadog.yaml.