pkg/trace/api: limit concurrent otlp requests to prevent RAM exhaustion
By default, a gRPC server will allow practically unlimited concurrent RPC
requests, spawning a goroutine for each. The routine will read and
deserialize the payload before calling the handler.

This means that our memory usage is not reasonably bounded, since we
could be holding hundreds of payloads in memory waiting to be processed.

This commit limits the number of simultaneous RPC requests, forwarding
the backpressure from the pipeline to the client, and limiting our memory
to O(MaxConnections * MaxStreams * MaxPayloadSize) in the receiver, with
MaxStreams being 1.