Always filter nil value from FX groups
Nil value can be inserted in FX groups when a component declare
returning a group element in the return types from its constructor but
leave the element to nil. This can happen with component that can be
disabled or partially enabled.

Filtering the FX group in the constructor or component using them
prevent panics down the line.