usm: http2: Introduce probe to clean http2 dynamic table
We introduce a new pair of probes per connenction (one for plaintext and one for TLS) to handle the cleanup
of the dynamic table.

We maintain 2 pointers per connection:
  1. "current" - dynamic table counter (the existing value we have).
  2. "previous" - used to store the latest entry we've clean.

After processing the headers, the cleanup phase kicks off. If "current" - "previous" reaches the threshold,
 we're cleaning all 'evicted' entries.

HTTP2 dynamic table is 4KB by default, but subject to changes per connection. Each entry consists of
  - the length of the key
  - the length of the value
  - and a constant of 32B.
(source - datatracker.ietf.org/doc/html/rfc7541#section-4.1)
If we assume all keys and values have a length of 1B, then we get a lower bound of 34B per
entry, which means we can store up to 120 entries (4KB / 32B) in the dynamic table before the eviction
happens.

Caveats:
  1. Since we don't really know what entries have been inserted to the dynamic table, we assume all
     of them were inserted, and we try to delete every entry in a specific range, hence there is at least 1 map
     operation overhead.
  2. We set a threshold of 200 entries, to support cases where the dynamic table size is bigger
     than 4KB. That means if we have more than 200 entries, any subsequent entry after 200th will be cleaned.
     For example, having the dynamic global counter pointing to 205, means we'll clean the entries 1 to 5.
  3. While we set the threshold for triggering the cleanup to 200, we actually can delete up to
     300 entries, to cover edge cases and gaps. For example in case a new frame introduces multiple new literal
     headers, and we cross the 200 threshold. For example, the difference is 198 entries, and a new frame
     introduces 5 new entries, thus the diff now is 203, which triggers the cleanup but misses 3 evicted entries.