[WKINT-388][Windows][CWS] Add ddprocmon to crash list
This change adds ddprocmon to the list of processes that are monitored for crashes by
the agent crash detection module; which notifies us if one of our drivers crashes.

Testing Instructions
- Install agent
- Enable _only_ CWS
  - do not enable NPM or USM, because that will invalidate the test
- start the agent
- check the log for the agent crash detection module.  You should see
```
2024-02-22 15:51:08 PST | CORE | DEBUG | (pkg/collector/worker/check_logger.go:44 in CheckStarted) | check:agentcrashdetect | Running check...
2024-02-22 15:51:08 PST | CORE | DEBUG | (pkg/collector/worker/check_logger.go:61 in CheckFinished) | check:agentcrashdetect | Done running check
```
you should NOT see
```
2024-02-22 17:09:41 PST | CORE | INFO | (comp/checks/agentcrashdetect/agentcrashdetectimpl/agentcrashdetect.go:128 in Configure) | Agent Crash Detection module will not run; no required components running
```

For additional testing.  install and load the crash driver (location here).
It will crash.  Upon reboot, a crash detection should be generated.  If your host is in `us-1`, then a slack notification should appear in #windows-kernel-intergrations-ops