[WKINT-375][Windows][CWS] Normalize reporting of registry keys.
ETW returns the registry path in the format
\REGISTRY\MACHINE

Users are used to seeing HKEY_LOCAL_MACHINE

translate the ETW path to the expected path

This PR includes adding a new functional test for the registry notifications.

This PR also includes fixing a race condition in which the event object was
shared between two goroutines, causing it to get intermittently stomped on.