pkg/trace: detect unconfigured ContainerTags and mute logs
We should not emit errors when failing to resolve container tags if the
container tags resolver function has not been configured.

This commit detects when the function is not configured and returns
success instead of an error.

It also changes a log from error to debug, since this error may occur
during normal functioning and the log will be useful only when debugging
issues such as missing container tags.