pkg/trace/writer: always include error in log when dropping a payload
We don't log every error on sending in the sender, to avoid filling the
log with repeated messages. However, we should log the send failure error
whenever we're dropping a payload.

This commit causes the send error to be logged along with the drop
warning any time a payload is dropped due to a send failure.