[WKINT=396] Fix tests that are hanging in CI
There are two problems which may be fixed by this.

The previous version, the tests are inexplicably hanging (more work to do).
In this version, etw was broken, which was obscured by ignoring failed tests due
to the known hanging problem.

This PR fixes the fact that USM was broken, and may fix the hanging problem, too.