Fixes problem in underlying ETW code that was causing tests to intermittently fail.
Underlying problem is that the ETW message that was being used to
clean up a connection sometimes comes out of the expected order (
the close comes before the connection sent notificiation).  This was
causing us to intermittently drop a connection.  In the real world,
it would mean that every now and again a connection would be missed.

In the test, when that one case came up, the test would fail and
hang forever.

This PR fixes the underlying problem by switching the ETW notification that causes cleanup to a different message, which (empirically) always comes last.

Also add additional test that floods the web server.  With the 10k connection test, (empirically again) I always got the out of order at least once.
Of course, since it's not predictable, it's possible to do 10K conns and always get the original behavior.  But the change in the test should cover the out-of-order case and prove that it's fixed.