(fleet) fix the install path override
This PR fixes the override of the InstallPath variable when the agent is
built with a non-default installation dir.

This fixes running the updater on a system that has no
`/opt/datadog-agent/`

This isn't covered by tests right now because it's a bit tricky to do so
but it will naturally be covered soon as we introduce some e2e tests on
the updater.

commit-id:3c275e37