(fleet) multi-package state tracking (#23450)
* (fleet) multi-package state tracking

This PR adds multi-package state tracking.

I had to resort to using a `context.Context` to keep the remote request data we want to include as part of the state reported to RC. The pattern is an expected use of contexts but I would rather have keept this more simple to be honest.

The alternative was to pass around all the request state in the top level method of the updater. This was problematic for the local API and the local bootstrap were we'd have to "fake" requests and I was worried this would potentially lead to confusing the backend.

commit-id:8951cff1

* (fleet) multi-package state tracking

This PR adds multi-package state tracking.

I had to resort to using a `context.Context` to keep the remote request data we want to include as part of the state reported to RC. The pattern is an expected use of contexts but I would rather have keept this more simple to be honest.

The alternative was to pass around all the request state in the top level method of the updater. This was problematic for the local API and the local bootstrap were we'd have to "fake" requests and I was worried this would potentially lead to confusing the backend.

commit-id:8951cff1