[SBOM] Disable HistoryDockerfile analyzer
This change disables the HistoryDockerfile analyzer,

This misconfiguration analyzer was added in Trivy on 2023-01-30.
This was part of the recent bump of Trivy to version 2024-02-28.

However, misconfiguration analyzers were disabled in our version
of Trivy, because they depend on the trivy-iac library, which
allocates very large structures at init time.