[agentless-scanner] Implement ScannerConfig structure
This change adds a new ScannerConfig structure, which contains
the configuration of the agentless-scanner. This structure
is passed to all the functions that requires access to the
agentless-scanner configuration.

Since the ScannerConfig is filled in the main package, there
is no need to import pkg/config anymore in any of the other
packages.