(rc) use the director repository to download TUF targets
This PR uses the director repository instead of the config repository to
get targets when clients ask for them. Functionally there should be no
change but some TUF internals make the director repo much more
efficient due to the absence of delegations.

This is a followup of
https://github.com/DataDog/datadog-agent/pull/23811. That first PR made
the verification much more efficient, potentially leading to the RC
service being more available to local tracers leading to increased
number of queries processed by the agent. This PR is to avoid a
"regression" on that front.

commit-id:6ce21376