usm: http2: classification: Fixed flakiness
The test 'http2 traffic without grpc' was falky with an error - 'Post "http://2.2.2.2:9090": dial tcp 2.2.2.2:9090: connect: connection refused'
Since that's the only falky test from the suite, it is likely there is a race between starting the server (via goroutine) to running the test.
A simple solution would be to first start http2 server, and then start gRPC server (instead the opposite order today), to 'gain' more time
allowing the http2 server to start properly, without having the need to sleep or to implement 'wait for server' method.