Refactor scheduling of jmxfetch checks
Fixes race condition when jmxfetch would lose configurations and not
run the checks that should be running.

When a check is quickly re-scheduled with the same configuration (and
thus id) after being unscheduled, the runner may attempt running a new
instance before previous have terminated. This causes the runner to
skip it's only attempt at running the new check instance.

All that jmx checks in the agent do is manage entries in the config
map for jmxfetch. This can be done much more reliably by updating
hashmap directly from autoconfig events, rather than asynchronously
running gouroutines.

This patch adds a separate autodiscovery scheduler to do that.