serverless cold start perf: enlarge sample & print
Given the importance of cold start latency to Serverless Agent
performance, gathering a larger data set would enable us to better
estimate quantile statistics, as well as the accuracy of our estimate
(e.g., using a confidence interval).

This commit increases the iteration count of the cold start container
loop from 10 to 301 so that:

- this test still runs in an hour or less (hopefully)

- we can plot a histogram of the raw data to better understand the
  distribution of cold start times

- the data set is large enough to use bootstrapping methods to
  construct confidence intervals on the median

In addition, this commit should print the raw cold start data in a
CSV-like format on a single line so it's easy to import into a `numpy`
array for exploratory data analysis.

Larger data sets (like 1,000+) would be better, resources
permitting. Under some mild assumptions[^1], as the sample size
approaches infinity, the distribution of the median approaches a
normal distribution[^2], which can be used to compute confidence
intervals[^3].

[^1]: Assuming the cold start latency could be modeled as an
absolutely continuous random variable.

[^2]: Google "asymptotic normality median", or see chapter 21 of
A. W. van der Vaart, *Asymptotic Statistics*, Cambridge University
Press, 1998.

[^3]: In practice, one would use bootstrapping to approximate the
distribution of medians, which is sort of like computing a correction
to this asymptotic result. T. Hesterberg, "What teachers should know
about the bootstrap: resampling in the undergraduate statistics
curriculum", [arXiv:1411.5279](https://arxiv.org/abs/1411.5279) is a
nice introduction to why this approach is useful. Rigorous
mathematical justification for this argument can be found in Chapter 3
of P. Hall, *The Bootstrap ane Edgeworth Expansion*, Springer,
1992. The "correction" I allude to is the second (and sometimes third)
term in an Edgeworth expansion about a normal distribution.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>