serverless perf: give up on CSV when printing data
When printing the raw data, commas are inessential and space-delimited
raw data is better than no raw data at all. This commit removes the
`IFS` part of the line that prints the raw data to console because it
wasn't fit for purpose, and no commas were printed. I suspect I could
make this work if I really wanted to spend the time, but I don't think
it's worth doing so.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>