.golangci.yml: allow use of sync/atomic
We introduced this depcheck in #12736, in order to prevent use of
sync/atomic. The reasons for this can be read about in that ticket, but
in short, sync/atomic had issues with alignment and an API which was not
intuitive, allowing non-atomic access to atomic values.

Since go1.19, new types have been added to sync/atomic which are
auto-aligned and prohibit non-atomic access, similar to the
go.uber.org/atomic library we replaced sync/atomic with.

Now that sync/atomic is suitable for use, we should allow its use.
Eventually we may wish to migrate away from go.uber.org/atomic entirely,
but allowing sync/atomic is a good first step.