mark /comp/README.md as un-owned
Currently all new comps implicitely need a review from ASC, even if created in existing domains owned by other teams.
This comes from the requirement to edit /comp/README.md which is owned by ASC. This PR proposes to mark that specifc file as un-owned to avoid that bottleneck. If this is intended and desirable it's also fine to reject it.