[windows][cws] update windows etw logging helpers. (#24711)
Many ETW notifications share the same underlying structure;
we use a typecast to handle this.  However, this would result
in the log outputting the type as the first one implemented,
rather than the actual type of the event.  This change updates
the String() method to accurately reflect which actual type is
being printed.