Fixes problem with rename detection, and clean up tests.
Discovered that we weren't actually receiving the renamePath event
(and the fsctl notification is a thing on win11, but at least not on 2016)

This led to proper processing of the rename event, which is simpler for us to process anyway.

Cleaned up test routines to be able to process until a given event, or until
the opens/closes match.

Clean up registry test to not process file operations at all.