[windows][cws][wkint-434] Add new ETW events for file based processing. (#24756)
* add type conversions for notifications that have same type as 'setInformatoin'

* add read/write notifications (same underlying type)

* add deletePathArgs, and other notifications that use the same underlying type

* add nameDeleteArgs, and other notifications that use the same underlying type

* add appropriate case statements to actually catch new notifications we're interested in

* add local processing tests

* fix linting

* Fixes problem with rename detection, and clean up tests.

Discovered that we weren't actually receiving the renamePath event
(and the fsctl notification is a thing on win11, but at least not on 2016)

This led to proper processing of the rename event, which is simpler for us to process anyway.

Cleaned up test routines to be able to process until a given event, or until
the opens/closes match.

Clean up registry test to not process file operations at all.

* Add old & new path to rename event.

Remove accidental leftover code limiting the processing to our own pid

* remove copy/paste mistake made in splitting out tests