[windows][cws][wkint-435]  Add permissions change notifications.
This PR consumes the well known ETW provider `Microsoft-Windows-Security-Auditing` to listen for permissions change events.

This PR does _not_ include the CWS processing, it simply surfaces the notifications from the ETW provider.

Several notes:
- The process must be run as local system (i.e. as a service) or it will fail silently.  This is a Windows limitation.
- Auditing must be enabled on the filesystem
- Auditing must be enabled in Group Policy
- Currently, audit events are logged at `info`.  This is too verbose; it is for clarity/development purposes.  Should be moved to `trace` once the CWS component is complete.

Testing:
Automated tests, analogous to what's provided for file and registry events, are not included in this PR because the
current test framework does not have the ability to run the tests as a service.

Tests will be added when the full CWS implementation is complete.