omnibus cache: ignore more environment variables (#24842)
* omnibus cache: ignore PACKAGE_ARCH

the job name is part of the cache URL so we don't risk to mixup various
caches that way.

* omnibus cache: ignore PIP_INDEX_URL

python integrations are systematically rebuilt so this variable bears no
importance for our usecase.
However since the python cache isn't always activated, this can cause
cache misses

* omnibus cache: ignore RELEASE_VERSION*

* omnibus cache: ignore GITHUB_* prefix

* omnibus cache: don't display the diff since last omnibus change

* omnibus cache: ignore RC_* env variables

This will allow RC pipelines to leverage the cache as well

* omnibus cache: currate the env for macOS runners