[windows] [NPM] Fix shutdown deadlock.
Testing revealed an intermittent deadlock in the windows network tracer shutdown code.
There was a race condition in which the shutdown code would clear the event
handle, but it would be reused, leading to a hang on shutdown.

Automated testing already included (as it was automated testing that illuminated this problem)