regression tests: disable file_tree test
The `file_tree` test seems to be generating enough data that it times
out the services analyzing that data, and these timeouts affect a
significant number of Agent pull requests.

We do not want a single experiment to cause every pull request to fail
the regression detector, so this commit disables that experiment for
triage purposes, until we figure out next steps to return it to proper
functioning.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>