disabled file_tree experiment: mark erratic
Until we figure out why the `file_tree` regression detector experiment
is misbehaving, it seems prudent to mark it erratic, so that
re-enabling it does not cause detected regressions to fail our
continuous integration testing.

Eventually, we would like to change the `erratic` setting for this
experiment to `false`; setting it to `true` in this commit gives us
some time to investigate the causes of `file_tree` time outs without
impeding other work on Agent.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>