[SMP] Temporarily disable `file_tree` experiment in regression tests (#24932)
* regression tests: disable file_tree test

The `file_tree` test seems to be generating enough data that it times
out the services analyzing that data, and these timeouts affect a
significant number of Agent pull requests.

We do not want a single experiment to cause every pull request to fail
the regression detector, so this commit disables that experiment for
triage purposes, until we figure out next steps to return it to proper
functioning.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* disabled file_tree experiment: mark erratic

Until we figure out why the `file_tree` regression detector experiment
is misbehaving, it seems prudent to mark it erratic, so that
re-enabling it does not cause detected regressions to fail our
continuous integration testing.

Eventually, we would like to change the `erratic` setting for this
experiment to `false`; setting it to `true` in this commit gives us
some time to investigate the causes of `file_tree` time outs without
impeding other work on Agent.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

---------

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>