usm: testutil: Add control messages to unix proxy
Precise control over the sizes of the read/writes performed is needed to
test packet split handling.  Add control messages to the unix proxy
which send the size of the read/write to be performed.  The default mode
of the proxy is still to be transparent; the new control messages will
initially only be used from the upcoming Kafka tests.

These control messages should allow avoiding having to add a sleep
between write calls (as present in the HTTP/2 tests).