usm: kafka: Run tests with and without TLS
Add support for running the tests with TLS too.  In order to do this,
convert the tests to use the unix proxy since the unit test binary
itself is built without symbols and cannot be hooked.  For simplicity,
the unix proxy is used both in the cases with and without TLS.