usm: kafka: Introduce pktbuf abstraction
In order to avoid copy-pasting the code to support TLS, add an
abstraction called pktbuf which wraps the helpers necessary to access
the packet and the metadata depending on whether it is backed by an SKB
or a user space TLS buffer in a unified interface.

This abstraction is intended to be zero-cost.  We rely on the compiler
to optimize away the unused code paths and structure passing.  The
kafka_response_parser program is fully identical before and after this
change.  The kafka_filter program sees a growth of 7 bytes (not 100%
sure why).