[SMP] Re-enable `file_tree` experiment in regression tests (reverts #24932) (#25091)
* regression tests: re-enable file_tree experiment

The Single Machine Performance team believes we have resolved the
issue that was causing `file_tree` experiments to fail, so this commit
re-enables the `file_tree` experiment in regression tests.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* `file_tree` test: set `erratic` to false

The `file_tree` experiment is supposed to test `system-probe`, and we
would like to detect potential performance regressions for that
target. In order to do so, we need to set the `file_tree` test's
`erratic` setting to `false` so that the results of that
test/experiment are used to determine whether Agent performance
regressions have occurred. This commit makes that change in the
`file_tree` experiment's configuration.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

---------

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>