usm: protocols: Exclude unused maps
Instead of editing unused maps to have only 1 entry, we use the ExcludeMap option, which
does not create the map at all.
A nice side effect, since each map allocates at least 1 page, it should spare up to 120KB
(30 maps, each is 4KB when disabled) from the RSS of the agent.

This is a preliminary change that is required for the purpose of making USM protocols
runtime dynamic